LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/signal-abort-window-stop-in-onnavigate-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: AbortError: Navigation aborted

PASS window.stop() signals event.signal inside a navigate event handler
